18Carolinas / 1860: generalize buy_sell_par_shares
